+++ next.config.mjs
@@ -85,3 +85,5 @@ const nextConfig = {
 }
 
 export default nextConfig
+
+import('@opennextjs/cloudflare').then(m => m.initOpenNextCloudflareForDev());